Imports at the top again:
This finally removes the star imports from records files without adding
to warnings. It also seems the most satisfactory solution as more
explicit. We explicitly import the records we need from parent game.
Only caveat is that the import order of parent vs child game record files
matters, but that on the other hand there is an inherent complexity here
that would manifest somehow - at least it is now explicitly mirrored
in the import order.